my-fastapi-nginx/app1/requirements.txt

fastapi
uvicorn
openai-whisper
gradio
